feat: sector tracking
