+++ Client/Assets/Resources/Prefabs/Map/Map.prefab
@@ -3418,10 +3418,6 @@ PrefabInstance:
       propertyPath: m_Name
       value: 2F_Corridor_A
       objectReference: {fileID: 0}
-    - target: {fileID: 5437356930162733777, guid: 1772b9601cec99f4b9ced76f070ea5c2, type: 3}
-      propertyPath: m_Layer
-      value: 11
-      objectReference: {fileID: 0}
     - target: {fileID: 7030881829091506616, guid: 1772b9601cec99f4b9ced76f070ea5c2, type: 3}
       propertyPath: m_LocalPosition.x
       value: -67.75
