+++ Client/Assets/Resources/Prefabs/Map/SectorEntrance.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cac5023e26240ff4eb213c3a515f2125
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
