+++ Client/Assets/Resources/Prefabs/UI/Scene/UI_AlienIngame.prefab
@@ -32,6 +32,7 @@ RectTransform:
   m_LocalScale: {x: 0, y: 0, z: 0}
   m_ConstrainProportionsScale: 0
   m_Children:
+  - {fileID: 900606130089702867}
   - {fileID: 277740841914005403}
   - {fileID: 8967625498928899649}
   - {fileID: 2118461260201502109}
@@ -220,6 +221,108 @@ PrefabInstance:
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 2819248160265837892, guid: 42404a70a901c9e42bfccd821b84e2f8, type: 3}
   m_PrefabInstance: {fileID: 2649594948293716703}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &3125607831451510935
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 6576529410540298847}
+    m_Modifications:
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_Pivot.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_Pivot.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_AnchorMax.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_AnchorMax.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_AnchorMin.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_AnchorMin.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_SizeDelta.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5535641532313121958, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+      propertyPath: m_Name
+      value: UI_CurrentSector
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+--- !u!224 &900606130089702867 stripped
+RectTransform:
+  m_CorrespondingSourceObject: {fileID: 2819248160265837892, guid: f4fc435ce66a331409e18366c360e9b9, type: 3}
+  m_PrefabInstance: {fileID: 3125607831451510935}
   m_PrefabAsset: {fileID: 0}
 --- !u!1001 &3198614262795951832
 PrefabInstance:
