+++ Client/Assets/Resources/Prefabs/UI/SubItem/UI_CurrentSector.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f4fc435ce66a331409e18366c360e9b9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
