+++ Client/Assets/Scripts/Contents/Map/Sector.cs
@@ -7,8 +7,9 @@ using UnityEngine.Serialization;
 
 public class Sector : MonoBehaviour
 {
-    [SerializeField]
-    public Define.SectorName sectorName;
+    [SerializeField] // 인스펙터 창에서 지정을 위해 필요
+    private Define.SectorName _sectorName;
+    public Define.SectorName SectorName => _sectorName;
 
     private List<Transform> _itemSpawnPoints = new();
     private Transform _itemParent;
@@ -31,7 +32,6 @@ public class Sector : MonoBehaviour
         Transform spawnPoint = _itemSpawnPoints[Random.Range(0, _itemSpawnPoints.Count)];
 
         NetworkObject no = Managers.NetworkMng.Runner.Spawn(go, spawnPoint.position, spawnPoint.rotation);
-        //no.transform.SetParent(Managers.GameMng.MapSystem.gameObject.transform);
         no.transform.SetParent(_itemParent);
 
         BaseItemObject item = no.GetComponent<BaseItemObject>();
@@ -42,5 +42,23 @@ public class Sector : MonoBehaviour
         _itemSpawnPoints.Remove(spawnPoint);
 
         return true;
+    }
+
+    public void OnCreatureEnter(Creature creature)
+    {
+        Debug.Log($"{SectorName}: OnCreatureEnter");
+        if (creature is Crew)
+        {
+            // TODO: 전역효과 효과 적용 등
+        }
+    }
+
+    public void OnCreatureExit(Creature creature)
+    {
+        Debug.Log($"{SectorName}: OnCreatureExit");
+        if (creature is Crew)
+        {
+            // TODO: 전역효과 효과 해제 등
+        }
     }
 }
