+++ Client/Assets/Scripts/Contents/Map/SectorEntrance.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd44deb43c680c24c93655e774f82b53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
