+++ Client/Assets/Scripts/Systems/ItemSpawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffc228bebe0243e459c3ada1e2be34bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
