+++ Client/Assets/Scripts/UI/SubItem/UI_CurrentSector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3f235a713455eee4e97aa8d37b831518
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
